+++ src/clients/api/queries/useGetVaults/useGetVestingVaults/useGetXvsVaultPools.ts
@@ -10,7 +10,7 @@ import {
   getXvsVaultUserInfo,
   IGetXvsVaultUserInfoOutput,
 } from 'clients/api';
-import { XVS_TOKEN_ADDRESS } from './constants';
+import { XVS_TOKEN_ADDRESS } from '../constants';
 
 export interface IUseGetXvsVaultPoolsInput {
   poolsCount: number;
